Add views folder and move views components to it
